moved stores to own package
