+++ module-array/src/main/java/ch/hsr/adv/ui/array/logic/StepEventController.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.ui.array.logic;
 
-import ch.hsr.adv.ui.core.logic.ADVEvent;
-import ch.hsr.adv.ui.core.logic.EventManager;
+import ch.hsr.adv.ui.core.logic.events.ADVEvent;
+import ch.hsr.adv.ui.core.logic.events.EventManager;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import org.slf4j.Logger;
