+++ ui-core/src/main/java/ch/hsr/adv/ui/core/logic/ADVFlowControl.java
@@ -6,6 +6,8 @@ import ch.hsr.adv.ui.core.logic.domain.Session;
 import ch.hsr.adv.ui.core.logic.domain.Snapshot;
 import ch.hsr.adv.ui.core.logic.events.ADVEvent;
 import ch.hsr.adv.ui.core.logic.events.EventManager;
+import ch.hsr.adv.ui.core.logic.stores.LayoutedSnapshotStore;
+import ch.hsr.adv.ui.core.logic.stores.SessionStore;
 import ch.hsr.adv.ui.core.logic.util.ADVParseException;
 import ch.hsr.adv.ui.core.presentation.util.I18n;
 import com.google.inject.Inject;
