+++ ui-core/src/main/java/ch/hsr/adv/ui/core/logic/stores/LayoutedSnapshotStore.java
@@ -1,4 +1,4 @@
-package ch.hsr.adv.ui.core.logic;
+package ch.hsr.adv.ui.core.logic.stores;
 
 import ch.hsr.adv.ui.core.logic.domain.LayoutedSnapshot;
 import ch.hsr.adv.ui.core.logic.events.ADVEvent;
