+++ ui-core/src/main/java/ch/hsr/adv/ui/core/presentation/RootViewModel.java
@@ -2,11 +2,11 @@ package ch.hsr.adv.ui.core.presentation;
 
 import ch.hsr.adv.ui.core.access.DatastoreAccess;
 import ch.hsr.adv.ui.core.logic.FlowControl;
-import ch.hsr.adv.ui.core.logic.LayoutedSnapshotStore;
-import ch.hsr.adv.ui.core.logic.SessionStore;
 import ch.hsr.adv.ui.core.logic.domain.Session;
 import ch.hsr.adv.ui.core.logic.events.ADVEvent;
 import ch.hsr.adv.ui.core.logic.events.EventManager;
+import ch.hsr.adv.ui.core.logic.stores.LayoutedSnapshotStore;
+import ch.hsr.adv.ui.core.logic.stores.SessionStore;
 import ch.hsr.adv.ui.core.presentation.util.I18n;
 import javafx.application.Platform;
 import javafx.beans.property.*;
