+++ ui-core/src/test/java/ch/hsr/adv/ui/core/logic/ADVFlowControlTest.java
@@ -6,6 +6,9 @@ import ch.hsr.adv.ui.core.logic.mocks.GuiceTestModule;
 import ch.hsr.adv.ui.core.logic.mocks.TestLayouter;
 import ch.hsr.adv.ui.core.logic.mocks.TestSession;
 import ch.hsr.adv.ui.core.logic.mocks.TestStringifyer;
+import ch.hsr.adv.ui.core.logic.stores.LayoutedSnapshotStore;
+import ch.hsr.adv.ui.core.logic.stores.SessionStore;
+import ch.hsr.adv.ui.core.logic.stores.SessionStoreTest;
 import ch.hsr.adv.ui.core.presentation.GuiceCoreModule;
 import com.google.inject.Inject;
 import org.jukito.JukitoRunner;
